Moved CANTalonSRX PIDF scaling to annonymous class method call(UNTESTED)
Add setPIDF method to CANTalonSRX that can be overrided to allow instances
to change the scaling between the map PIDF and the Talon input PIDF

Override the setPIDF method where the flywheel and turntable Talons are
constructed, hardcoding the scalings in private anonymous class methods.

TODO: Move the hardcoded scalings to the flywheel and turntable
maps.